backend/backend/requirements.txt
# FILE: backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
numpy==2.1.1
opencv-python==4.10.0.84
Pillow==10.4.0
scikit-image==0.24.0
pyinstaller==6.10.0
requests==2.32.3
openai==1.68.2
